chore: migrate REST policy classes to Kotlin
