+++ src/main/java/net/atos/zac/app/informatieobjecten/model/RestEnkelvoudigInformatieobject.java
@@ -13,9 +13,9 @@ import java.util.UUID;
 import jakarta.validation.constraints.NotNull;
 import jakarta.ws.rs.FormParam;
 
-import net.atos.zac.app.policy.model.RestDocumentRechten;
 import nl.info.client.zgw.drc.model.generated.StatusEnum;
 import nl.info.zac.app.identity.model.RestUser;
+import nl.info.zac.app.policy.model.RestDocumentRechten;
 import nl.info.zac.search.model.DocumentIndicatie;
 
 /**
