+++ src/main/kotlin/nl/info/zac/app/policy/model/RestTaakRechten.kt
@@ -1,9 +1,10 @@
 /*
- * SPDX-FileCopyrightText: 2022 Atos
+ * SPDX-FileCopyrightText: 2022 Atos, 2025 INFO.nl
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.zac.app.policy.model
+package nl.info.zac.app.policy.model
 
+import nl.info.zac.policy.output.TaakRechten
 import nl.info.zac.util.NoArgConstructor
 
 // No-arg constructor is needed because somewhere we
@@ -15,4 +16,11 @@ data class RestTaakRechten(
     val wijzigen: Boolean,
     val toekennen: Boolean,
     val toevoegenDocument: Boolean
+)
+
+fun TaakRechten.toRestTaakRechten() = RestTaakRechten(
+    lezen = this.lezen,
+    wijzigen = this.wijzigen,
+    toekennen = this.toekennen,
+    toevoegenDocument = this.toevoegenDocument
 )
