+++ src/main/kotlin/nl/info/zac/app/zaak/converter/RestGerelateerdeZaakConverter.kt
@@ -9,8 +9,8 @@ import net.atos.client.zgw.zrc.ZrcClientService
 import net.atos.client.zgw.zrc.model.AardRelatie
 import net.atos.client.zgw.zrc.model.RelevanteZaak
 import net.atos.client.zgw.zrc.model.Zaak
-import net.atos.zac.app.policy.converter.toRestZaakRechten
 import nl.info.client.zgw.ztc.ZtcClientService
+import nl.info.zac.app.policy.model.toRestZaakRechten
 import nl.info.zac.app.zaak.model.RelatieType
 import nl.info.zac.app.zaak.model.RestGerelateerdeZaak
 import nl.info.zac.policy.PolicyService
