+++ src/main/kotlin/nl/info/zac/app/zaak/converter/RestZaakOverzichtConverter.kt
@@ -7,13 +7,13 @@ package nl.info.zac.app.zaak.converter
 import jakarta.inject.Inject
 import net.atos.client.zgw.zrc.ZrcClientService
 import net.atos.client.zgw.zrc.model.Zaak
-import net.atos.zac.app.policy.converter.toRestZaakRechten
 import nl.info.client.zgw.shared.ZGWApiService
 import nl.info.client.zgw.ztc.ZtcClientService
 import nl.info.zac.app.identity.converter.RestGroupConverter
 import nl.info.zac.app.identity.converter.RestUserConverter
 import nl.info.zac.app.identity.model.RestGroup
 import nl.info.zac.app.identity.model.RestUser
+import nl.info.zac.app.policy.model.toRestZaakRechten
 import nl.info.zac.app.zaak.model.RestZaakOverzicht
 import nl.info.zac.policy.PolicyService
