+++ src/main/kotlin/nl/info/zac/app/zaak/model/RestZaakOverzicht.kt
@@ -5,9 +5,9 @@
 package nl.info.zac.app.zaak.model
 
 import com.fasterxml.jackson.annotation.JsonFormat
-import net.atos.zac.app.policy.model.RestZaakRechten
 import nl.info.zac.app.identity.model.RestGroup
 import nl.info.zac.app.identity.model.RestUser
+import nl.info.zac.app.policy.model.RestZaakRechten
 import nl.info.zac.util.AllOpen
 import nl.info.zac.util.NoArgConstructor
 import java.time.LocalDate
