+++ src/test/kotlin/nl/info/zac/app/task/model/TaskRestFixtures.kt
@@ -9,6 +9,7 @@ import jakarta.json.JsonObject
 import net.atos.zac.app.informatieobjecten.model.RestInformatieobjecttype
 import nl.info.zac.app.identity.model.RestUser
 import nl.info.zac.app.informatieobjecten.model.createRestInformatieobjecttype
+import nl.info.zac.app.policy.model.createRestTaakRechten
 import nl.info.zac.app.zaak.model.createRestUser
 import java.util.UUID
 
@@ -28,7 +29,8 @@ fun createRestTask(
     tabellen = tabellen,
     zaakIdentificatie = zaakIdentificatie,
     zaakUuid = zaakUuid,
-    formioFormulier = formioFormulier
+    formioFormulier = formioFormulier,
+    rechten = createRestTaakRechten(),
 )
 
 fun createRestTaskDocumentData(
